Refactor telemetry UI boundaries
