+++ controller/src/telemetry/telemetry_live_store.cpp
@@ -2,13 +2,15 @@
 
 #include <utility>
 
+#include "telemetry/telemetry_live_store_services.h"
 #include "telemetry/telemetry_live_store_state.h"
 #include "telemetry/telemetry_state_types.h"
 
 namespace naim::controller {
 
 TelemetryLiveStore::TelemetryLiveStore()
-    : state_(std::make_unique<TelemetryLiveStoreState>()) {}
+    : state_(std::make_unique<TelemetryLiveStoreState>()),
+      services_(std::make_unique<TelemetryLiveStoreServices>()) {}
 
 TelemetryLiveStore::~TelemetryLiveStore() = default;
 
@@ -26,9 +28,9 @@ void TelemetryLiveStore::ConfigurePersistence(
 
 void TelemetryLiveStore::ConfigureFromEnvironment(const std::string& db_path) {
   const auto retention =
-      state_->operational_config.RetentionFromEnvironment(state_->retention);
+      services_->operational_config.RetentionFromEnvironment(state_->retention);
   const auto thresholds =
-      state_->operational_config.ThresholdsFromEnvironment(state_->thresholds);
+      services_->operational_config.ThresholdsFromEnvironment(state_->thresholds);
   ConfigureOperationalPolicy(retention, thresholds);
   ConfigurePersistence(db_path, retention.durable_history_capacity);
 }
@@ -37,9 +39,9 @@ void TelemetryLiveStore::ConfigureOperationalPolicy(
     RetentionConfig retention,
     AlertThresholds thresholds) {
   std::lock_guard<std::mutex> lock(mutex_);
-  state_->retention = state_->operational_config.NormalizeRetention(retention);
+  state_->retention = services_->operational_config.NormalizeRetention(retention);
   state_->thresholds = thresholds;
-  state_->ring_buffer.ApplyHotRetention(
+  services_->ring_buffer.ApplyHotRetention(
       state_->nodes,
       state_->dropped_frames_total,
       state_->retention);
@@ -60,11 +62,11 @@ bool TelemetryLiveStore::Upsert(naim::HostTelemetryFrame frame) {
   if (frame.node_name.empty()) {
     return false;
   }
-  frame = state_->frame_normalizer.Normalize(std::move(frame));
+  frame = services_->frame_normalizer.Normalize(std::move(frame));
   std::lock_guard<std::mutex> lock(mutex_);
   const bool updated = UpsertInMemoryLocked(frame);
   if (updated) {
-    state_->persistence_repository.PersistFrame(state_->persistence, frame);
+    services_->persistence_repository.PersistFrame(state_->persistence, frame);
   }
   return updated;
 }
@@ -73,7 +75,7 @@ nlohmann::json TelemetryLiveStore::BuildSnapshot(
     const std::optional<std::string>& plane_name,
     const int history_seconds) const {
   std::lock_guard<std::mutex> lock(mutex_);
-  return state_->snapshot_builder.BuildSnapshot(
+  return services_->snapshot_builder.BuildSnapshot(
       state_->nodes,
       state_->retention,
       state_->thresholds,
@@ -83,13 +85,13 @@ nlohmann::json TelemetryLiveStore::BuildSnapshot(
       state_->dropped_frames_total,
       plane_name,
       history_seconds,
-      state_->clock.CurrentUnixMillis());
+      services_->clock.CurrentUnixMillis());
 }
 
 nlohmann::json TelemetryLiveStore::BuildHealth(
     const std::optional<std::string>& plane_name) const {
   std::lock_guard<std::mutex> lock(mutex_);
-  return state_->health_builder.BuildHealth(
+  return services_->health_builder.BuildHealth(
       state_->nodes,
       state_->retention,
       state_->thresholds,
@@ -98,19 +100,19 @@ nlohmann::json TelemetryLiveStore::BuildHealth(
       state_->latest_sequence,
       state_->dropped_frames_total,
       plane_name,
-      state_->clock.CurrentUnixMillis());
+      services_->clock.CurrentUnixMillis());
 }
 
 std::string TelemetryLiveStore::BuildOpenMetrics(
     const std::optional<std::string>& plane_name) const {
-  return state_->open_metrics_exporter.Build(BuildHealth(plane_name), plane_name);
+  return services_->open_metrics_exporter.Build(BuildHealth(plane_name), plane_name);
 }
 
 std::vector<naim::HostTelemetryFrame> TelemetryLiveStore::LoadFramesAfter(
     const std::uint64_t sequence,
     const std::optional<std::string>& plane_name) const {
   std::lock_guard<std::mutex> lock(mutex_);
-  return state_->ring_buffer.LoadFramesAfter(
+  return services_->ring_buffer.LoadFramesAfter(
       state_->nodes,
       state_->retention,
       sequence,
@@ -121,7 +123,7 @@ TelemetryLiveStore::StreamDelta TelemetryLiveStore::LoadStreamDeltaAfter(
     const std::uint64_t sequence,
     const std::optional<std::string>& plane_name) const {
   std::lock_guard<std::mutex> lock(mutex_);
-  return state_->ring_buffer.LoadStreamDeltaAfter(
+  return services_->ring_buffer.LoadStreamDeltaAfter(
       state_->nodes,
       state_->retention,
       state_->latest_sequence,
@@ -137,26 +139,26 @@ std::uint64_t TelemetryLiveStore::LatestSequence() const {
 
 void TelemetryLiveStore::RecordStreamOpened(const std::string& stream_name) {
   std::lock_guard<std::mutex> lock(mutex_);
-  state_->stream_metrics_service.RecordOpened(state_->streams, stream_name);
+  services_->stream_metrics_service.RecordOpened(state_->streams, stream_name);
 }
 
 void TelemetryLiveStore::RecordStreamClosed(const std::string& stream_name) {
   std::lock_guard<std::mutex> lock(mutex_);
-  state_->stream_metrics_service.RecordClosed(state_->streams, stream_name);
+  services_->stream_metrics_service.RecordClosed(state_->streams, stream_name);
 }
 
 void TelemetryLiveStore::RecordStreamReplayRequired(const std::string& stream_name) {
   std::lock_guard<std::mutex> lock(mutex_);
-  state_->stream_metrics_service.RecordReplayRequired(state_->streams, stream_name);
+  services_->stream_metrics_service.RecordReplayRequired(state_->streams, stream_name);
 }
 
 void TelemetryLiveStore::RecordStreamSendFailure(const std::string& stream_name) {
   std::lock_guard<std::mutex> lock(mutex_);
-  state_->stream_metrics_service.RecordSendFailure(state_->streams, stream_name);
+  services_->stream_metrics_service.RecordSendFailure(state_->streams, stream_name);
 }
 
 bool TelemetryLiveStore::UpsertInMemoryLocked(naim::HostTelemetryFrame frame) {
-  return state_->ring_buffer.Upsert(
+  return services_->ring_buffer.Upsert(
       state_->nodes,
       state_->latest_sequence,
       state_->dropped_frames_total,
@@ -168,12 +170,12 @@ void TelemetryLiveStore::ConfigurePersistenceLocked(
     const std::string& db_path,
     const std::size_t retention_capacity) {
   const auto frames =
-      state_->persistence_repository.Configure(
+      services_->persistence_repository.Configure(
           state_->persistence,
           db_path,
           retention_capacity);
   for (auto frame : frames) {
-    frame = state_->frame_normalizer.Normalize(std::move(frame));
+    frame = services_->frame_normalizer.Normalize(std::move(frame));
     UpsertInMemoryLocked(std::move(frame));
   }
 }
