+++ controller/src/telemetry/telemetry_live_store_services.cpp
@@ -1,0 +1,7 @@
+#include "telemetry/telemetry_live_store_services.h"
+
+namespace naim::controller {
+
+TelemetryLiveStoreServices::TelemetryLiveStoreServices() = default;
+
+}  // namespace naim::controller
